+++ sdk-java/src/test/java/ly/count/sdk/java/internal/BackendModeTests.java
@@ -25,14 +25,14 @@ public class BackendModeTests {
         cc.setEventQueueSizeToSend(4).enableBackendMode();
 
         // System specific folder structure
-        String[] targetPath = { System.getProperty("user.home"), "__COUNTLY", "java_test" };
-        File targetFolder = new File(String.join(File.separator, targetPath));
+        String[] sdkStorageRootPath = { System.getProperty("user.home"), "__COUNTLY", "java_test" };
+        File sdkStorageRootDirectory = new File(String.join(File.separator, sdkStorageRootPath));
 
-        if(targetFolder.mkdirs()){
+        if(sdkStorageRootDirectory.mkdirs()){
             System.out.println("Directory creation failed");
         }
 
-        Countly.init(targetFolder, cc);
+        Countly.init(sdkStorageRootDirectory, cc);
     }
 
     @AfterClass